Finish question before 20;

+++ LeetCode/16.cpp
@@ -12,7 +12,7 @@ public:
     int threeSumClosest(vector<int>& nums, int target) {
         int minDiff = INT_MAX;
         int ans = 0;
-        int len = nums.size();
+        int len = (int)nums.size();
         int sum, l, r;
         sort(nums.begin(), nums.end());
         for (int i = 0; i < len; i++) {
